Portage Mingw et revision du portage Win32/MSVC
git-svn-id: http://caml.inria.fr/svn/ocaml/trunk@4899 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ otherlibs/win32unix/lseek.c
@@ -14,6 +14,7 @@
 /* $Id$ */
 
 #include <mlvalues.h>
+#include <alloc.h>
 #include "unixsupport.h"
 
 #ifdef HAS_UNISTD
@@ -71,5 +72,5 @@ CAMLprim value unix_lseek_64(value fd, value ofs, value cmd)
       uerror("lseek", Nothing);
     }
   }
-  return copy_int64(ofs_high << 32 | ret);
+  return copy_int64((int64) ofs_high << 32 | ret);
 }
